+++ test/e2e/tests/basic/x_enum_names/expected_freezed/models/status.dart
@@ -1,0 +1,34 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint, unused_import, invalid_annotation_target, unnecessary_import
+
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+@JsonEnum()
+enum Status {
+  @JsonValue(200)
+  ok(200),
+  @JsonValue(404)
+  notFound(404),
+  @JsonValue(500)
+  serverError(500),
+
+  /// Default value for all unparsed values, allows backward compatibility when adding new values on the backend.
+  $unknown(null);
+
+  const Status(this.json);
+
+  factory Status.fromJson(int json) =>
+      values.firstWhere((e) => e.json == json, orElse: () => $unknown);
+
+  final int? json;
+
+  String toJson() => json?.toString() ?? 'null';
+
+  @override
+  String toString() => json?.toString() ?? super.toString();
+
+  /// Returns all defined enum values excluding the $unknown value.
+  static List<Status> get $valuesDefined =>
+      values.where((value) => value != $unknown).toList();
+}